+++ resources/views/account/notifications.blade.php
@@ -1,0 +1,5 @@
+<x-layouts::storefront :title="__('Notifications')">
+    <x-account.layout>
+        <livewire:storefront.notifications-page />
+    </x-account.layout>
+</x-layouts::storefront>